+++ src/test/java/org/folio/indexing/IndexingLocationsConsortiumIT.java
@@ -2,6 +2,7 @@ package org.folio.indexing;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
+import static org.awaitility.Durations.ONE_HUNDRED_MILLISECONDS;
 import static org.awaitility.Durations.ONE_MINUTE;
 import static org.awaitility.Durations.ONE_SECOND;
 import static org.folio.search.domain.dto.ResourceEventType.CREATE;
@@ -10,16 +11,22 @@ import static org.folio.search.domain.dto.ResourceEventType.DELETE_ALL;
 import static org.folio.search.model.types.ResourceType.LOCATION;
 import static org.folio.support.TestConstants.CENTRAL_TENANT_ID;
 import static org.folio.support.TestConstants.MEMBER_TENANT_ID;
+import static org.folio.support.TestConstants.inventoryLibraryTopic;
 import static org.folio.support.TestConstants.inventoryLocationTopic;
 import static org.folio.support.utils.JsonTestUtils.toMap;
 import static org.folio.support.utils.TestUtils.kafkaResourceEvent;
 import static org.folio.support.utils.TestUtils.randomId;
+import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.Mockito.doAnswer;
+import static org.mockito.Mockito.reset;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicBoolean;
 import org.folio.search.domain.dto.Metadata;
 import org.folio.search.domain.dto.ResourceEvent;
+import org.folio.search.integration.message.KafkaMessageListener;
 import org.folio.search.model.dto.LocationDto;
 import org.folio.spring.testing.type.IntegrationTest;
 import org.folio.support.base.BaseConsortiumIntegrationTest;
@@ -27,9 +34,12 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 
 @IntegrationTest
 class IndexingLocationsConsortiumIT extends BaseConsortiumIntegrationTest {
+
+  @MockitoSpyBean private KafkaMessageListener kafkaMessageListener;
 
   @BeforeAll
   static void prepare() {
@@ -46,6 +56,7 @@ class IndexingLocationsConsortiumIT extends BaseConsortiumIntegrationTest {
   @AfterEach
   void tearDown() throws IOException {
     cleanUpIndex(LOCATION, CENTRAL_TENANT_ID);
+    reset(kafkaMessageListener);
   }
 
   @Test
@@ -82,6 +93,27 @@ class IndexingLocationsConsortiumIT extends BaseConsortiumIntegrationTest {
     var deleteAllMemberEvent = new ResourceEvent().type(DELETE_ALL).tenant(MEMBER_TENANT_ID);
     kafkaTemplate.send(inventoryLocationTopic(MEMBER_TENANT_ID), deleteAllMemberEvent);
     awaitAssertLocationCount(1);
+  }
+
+  @Test
+  void shouldIgnoreShadowInstitution() {
+    var libraryMap = toMap(location());
+    libraryMap.put("isShadow", true);
+    var visited = new AtomicBoolean(false);
+
+    doAnswer(inv -> {
+      visited.set(true);
+      return inv.callRealMethod();
+    }).when(kafkaMessageListener).handleLocationEvents(anyList());
+
+    var libraryEvent = kafkaResourceEvent(CENTRAL_TENANT_ID, CREATE, libraryMap, null);
+    kafkaTemplate.send(inventoryLibraryTopic(MEMBER_TENANT_ID), libraryEvent);
+
+    await().atMost(ONE_MINUTE)
+      .pollInterval(ONE_HUNDRED_MILLISECONDS)
+      .untilAsserted(() -> assertThat(visited.get()).isTrue());
+
+    awaitAssertLocationCount(0);
   }
 
   public static void awaitAssertLocationCount(int expected) {